MB-5259: ask memcached to reread isasl.pw when we update the file
Recent memcached change now requires us to send refresh-isasl command
for memcached to re-read this file. This is simplest possible change
to make new isasl work.

NOTE: new command allows us to be aware when memcached is done
re-reading isasl. Which will allow us to make bucket creation
synchronous more easily. But for now this is not a goal.

Change-Id: I7a588760fcb848ac66a05955251cea1d895ebb49
Reviewed-on: http://review.couchbase.org/16503
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>