MB-5307:implemented more efficient ebucketmigrator initialization
Particularly we're dealing with larger set of vbuckets, so we don't want
very chatty exchange with memcacheds for things like finding out
open/closed checkpoints and setting destination vbuckets to replica
state.

Particularly:

a) instead of doing a bunch of sequential set_vbucket requests we
first check which vbuckets needs to be set to replica state

b) instead of doing a bunch of sequential get_last_closed_checkpoint
requests we're doing single stats request. And we're using quick stats
to get this.

c) we're using quick stats for getting list of open checkpoints

Change-Id: I99b4bcfa74bd64974c306ee7ab386302595d0b05
Reviewed-on: http://review.couchbase.org/16780
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>