MB-5434: don't add ejected nodes back to bucket's servers list
It's possible that ejected node(s) is not part of bucket's node list
anymore or yet. One possibility is if it was already rebalanced out
for this bucket.

If that is the case we should not add it back to server's list at the
beginning of rebalance, which would cause unnecessary waiting for
bucket startup on this node.

Change-Id: I70e1ed622b5246aca0f32000d51f67bcc08c6da4
Reviewed-on: http://review.couchbase.org/16783
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>