MB-5307: fixed getting of not ready to replicate from vbuckets
Implementation of get_open_checkpoint_ids/2 was returning dict with
stat names as keys instead of vbucket numbers. That was causing us to
never detect any vbuckets as not ready to replicate from.

I decided to optimize things a bit further by asking specifically for
vbuckets with open checkpoint 0 and by trying to minimize produced
garbage that was typical for previous implementation that was building
dict.

Change-Id: Ice165d2161348bc9f92a99d8340a8ab0bc9f1f5e
Reviewed-on: http://review.couchbase.org/17081
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>