MB-5505: Make sure we do not over-split the VB list
Today when XDCR is scheduling replication, it splits the vbucket list
into two [list1, list2] where length of list1 is number of concurrent
replication streams defined in an env parameter. But there could
be a case that the parameter value is greater than the length of
vbucket list. This changeset will fix this issue by determing the
right position to split the vbucket list.

Change-Id: I32a84b8082114bc39e13e34338d9d9d882196080
Reviewed-on: http://review.couchbase.org/17169
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>