MB-5734 Return random keys efficiently.
In most common case when there's a lots of documents in a bucket we'll
be using couch_db:random_doc_info function. In case there are too few
documents, old _all_docs approach will be used (but on the client's
side).

Corresponding REST call moved to
/pools/default/buckets/<bucket>/localRandomKey.

On the UI side any caching of random keys is removed since we expect
this to be fast now.

Change-Id: I1a4331e912bc2ad0e3918e56f1513ff5549e9c6e
Reviewed-on: http://review.couchbase.org/17836
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>