CBD-399: refactoring checkpoint code in replicator
Today the replicator module consists of mixed code for replication
management and checkpointing, which is not neat and make replicator
module unnecessarily bloated. For better code style and easier future
debug and developement, this change-set creates a new module, namely
xdc_replicator_ckpt, and removes all checkpoint realated code to the
new module.

Change-Id: I72427e056da4a201e539c4c6afec571431119e78
Reviewed-on: http://review.couchbase.org/17899
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>