CBD-399: create new entry in stat table before creating replicators
Today we insert an entry into stat table after we create a replicator
process, which may subject to race condition that replicator process
is trying to query and update the stat table before its entry is
inserted into that table. This change-set fixes this issue by creating
an entry in stat table before creating the corresponding replicator
process.

Change-Id: I7980e41f485a3524ee2ac8c77be9549deda513f2
Reviewed-on: http://review.couchbase.org/18052
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>