CBD-378 A service to access remote clusters information.
Provides several functions to fetch information about remote clusters
and remote clusters' buckets. Whenever this information is inquired
it's cached locally. So that next queries could be served from this
cache. In case the client finds out the provided information to be
inconsistent it can force remote_cluster_info to go to the remote
cluster to get up-to-date data.

Change-Id: Ie3e27c992fd065151abcb977cfa20dff5f8460b5
Reviewed-on: http://review.couchbase.org/18831
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>