MB-5307: made ebucketmigrator_srv:had_backfill reply after ack
Previous implementation was incorrectly assuming that seeing
backfill-ness evidence message is enough. But at least of initial
stream message, which is positive evidence of backfill, it's possible
to have result of had_backfill before resetVbucket on destination is
really done. Which we don't want as we want 'stable' checkpoint ids in
destination after had_backfill returns.

Change-Id: Ibeb496b592f8668ff28c22d167dc3237572fbb15
Reviewed-on: http://review.couchbase.org/18797
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>