CBD-181 Advertise cancel URIs in tasks for manual compactions.
Note that at least currently URIs are advertised only for bucket or
databases compactions. That's because we distinguish between scheduled
and manual compactions by the trigger_type field in compaction
status. And we don't control view compactor's status. It's not a
problem right now because the only kind of manual compaction we
provide on our UI is bucket compaction. Something a little bit smarter
will have to be invented to support this.

Change-Id: Ie80d13ff9212d5a6deea44989008a3cc6a4a9ca3
Reviewed-on: http://review.couchbase.org/19119
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>