CBD-181: implemented compact button on views
As part of this change simple UI-service for keeping track of recently
compacted ddocs is introduced. It keeps track of compactions that were
started 10 seconds ago or later. So that we can avoid showing compact
button as enabled for 10 seconds after it's pressed.

The idea is that because there's no way right now to know if server is
starting or is going to start or actually completed previously
requested compaction, the only way to intelligently disable this
button is by keeping track of recently started compactions on UI.

Change-Id: Idfac0a29bf46f3fbfeb50e87612cb0550b1a24df
Reviewed-on: http://review.couchbase.org/19142
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>