CBD-423 Resurrect set_view_update_daemon in ns_server.
Configuration parameters were moved to ns_server config. Update
interval can only be specified globally for all design
documents. Minimum number of changes to update the index can be
specified globally and on per design document basis. For the latter
one should set "updateMinChanges" integer field in "options" object in
design document body. This value will override a global value.

updateMinChanges value of 0 (either global or per ddoc) means that
update won't be triggered for this ddoc (or for all ddocs). But the
daemon will still periodically try to trigger updates.

To disable daemon completely updateInterval should be set to 0.

For simplicity the code behaves as if there were no design
documents with the same signature. In reality if there happen to be
several design documents with the same signature, then update will be
triggered whenever at least one design document's updateMinChanges
value is less than the number of changes.

Change-Id: Idcc2c4e62bd76160f4475c407d1d06bfe6ec8e8d
Reviewed-on: http://review.couchbase.org/19260
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>