MB-5307 Rename subtype to original_target.
It seems to make more sense. The value of this status field indicates
the original target of compactions. For instance, database compaction
can be triggered directly (by user) of as a result of bucket
compaction (schedule or triggered by user as well). So the value
allows to distinguish these cases.

Change-Id: I4dfbbab7d083a6ff7ba68053433e7ccc36c52b46
Reviewed-on: http://review.couchbase.org/19266
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>