CBD-178: implemented index-aware vbucket moving
In order to have 'consistent views' during rebalance we need to make
sure indexes for vbucket being moved are at least as up to date on new
master as on old master.

The following is performed to achieve that:

* on replicas and future master we're waiting to persist at least same
  docs as were persisted on old master

* indexing on old master is paused for current vbucket to ensure index
  does not 'move forward' w.r.t. this vbucket

* we're waiting again while future master has persisted _and_ indexed
  (in passive state) at least same docs as were persisted on old
  master after pausing indexing.

* we perform vbucket takeover, switch vbucket into active state on new
  master and cleanup state on old master, then update vbucket
  map. From this point, queries will use vbucket's index on new
  master.

As can be seen, this implements 'monotonicity' of index from the point
of queries reading index.

New behavior is disable-able by ns_config variables:

* index_aware_rebalance_disabled

* index_pausing_disabled (just disabling pausing)

As of this commit both are disabled to enable fuller testing before
enabling them.

Change-Id: I732bb225673e8e8969747fc83fdc557b124b1a6c
Reviewed-on: http://review.couchbase.org/19347
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>