MB-5307 Don't actually remove remote cluster references.
Since it's not easy to reliably remove all the replications when
cluster reference removed (or alternatively, it's hard to disallow
removing cluster reference when there're still existent replications),
we decided not to remove cluster references. Instead they are marked
as deleted and are not shown on the UI.

We enforce two invariants on remote clusters lists:

  1. For every UUID there's only one cluster (both deleted and not
  deleted) with such UUID in the list.

  2. For any cluster name there's only one not deleted cluster with
  that name.

On the UI if remote cluster reference for some replication is deleted,
we show remote cluster's hostname instead of cluster reference
name. That's because when it comes to deleted references, the name is
not unique anymore.

Change-Id: I1c2540f288462275291b8bf37648f3f03842aebc
Reviewed-on: http://review.couchbase.org/19460
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>