MB-5895: consider upgrading vb map history at rebalance start
When 1.8.1 is added to cluster of 1.8.0 nodes there's initially no
vbucket map history. This happens because 1.8.1 currently considers
upgrading config only when offline upgrading. But for vbucket map
history which is global config element, we need to do that with actual
vbucket maps (after joining cluster) but before rebalance, so that
swap rebalance works.

Change-Id: I1757f869b5542b06ccf0951d8827e21963f1c0cc
Reviewed-on: http://review.couchbase.org/19387
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>