MB-5307 Handle duplicated alerts correctly.
Suppose some after some alert is sent for the first time. After
sending alert to all the nodes check_alerts handler will compute new
state based on the old state and will send that new state to the main
process. That new state may be delivered to the main process after the
alert already processed. And that case old history (actually empty
history) will overwrite the history that was updated by alert
handler. To prevent this new state is computed in the main process
based on current process state.

Change-Id: I7344cf2a03afaefdb27d9d98a439e49c6ef97108
Reviewed-on: http://review.couchbase.org/19745
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>