MB-6131 Send out email alerts asynchronously.
Although ns_mail_log is the only ns_log_events handler, it's still
doesn't make much sense to block in handle_event for a long time. As a
side-effect this commit introduces a way to send emails with
configurable timeout. It may be helpful because gen_smtp_client uses
hard-coded timeout of 1200 seconds.

Change-Id: Ic01d684a033056f408c2f8fe372379ee169ff6f7
Reviewed-on: http://review.couchbase.org/19744
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>