MB-6049: fixed race listing all present vbuckets
couch_server:all_databases is not really a race-free way to discover
bucket's vbuckets. Particularly because listing directory is not
atomic it's possible to not see some vbuckets that were compacted in
the middle of listing directory.

So we're using new couch_server:all_known_databases
in addition to all_databases in order to see all vbuckets.

This commit reverts revert of 100% same commit made because couchdb
dependency is not yet merged.

Change-Id: Ic3fe1967b6df68fcf1dce8e4696a8b31ec4ee3a9
Reviewed-on: http://review.couchbase.org/20089
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>