MB-6049: fixed race listing all present vbuckets
couch_server:all_databases is not really a race-free way to discover
bucket's vbuckets. Particularly because listing directory is not
atomic it's possible to not see some vbuckets that were compacted in
the middle of listing directory.

So we're using new couch_server:all_known_databases
in addition to all_databases in order to see all vbuckets.

Change-Id: Ia73eb8b7ce8d74df6c5e63e0417d485eeec323bd
Reviewed-on: http://review.couchbase.org/19753
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>