MB-6278 Ignore no_db_file errors caused by vbucket transfer.
Previously we would get many no_db_file crashes from compaction_daemon
if compaction was triggered right before rebalance resulting in many
vbuckets moved out of the node. To be more graceful here we'll consult
vbucket map when we encounter such an error. And if the vbucket has
been moved out of our node, we'll ignore this error.

Change-Id: Ie48144275623d71d6e17341c7303ec08b0b57e60
Reviewed-on: http://review.couchbase.org/19894
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>