MB-4454 --loglevel controls only what is sent to stderr.
Instead of setting all the log levels to provided value now it just
passes additional parameter to ns_server. This parameter specifies log
level only for stderr sink. Basically this means that from now
--loglevel will change only what's sent to stderr but not what is sent
to log files. So that it's possible to make stderr output less verbose
but in case of some issue refer to complete log files.

Change-Id: Ia8ac30c9dc56e241db19dfd22a1d485253c2c926
Reviewed-on: http://review.couchbase.org/20074
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>