MB-6468: force reading updated target vbucketmap
Today each vbucket replicator may read a stale target cluster vbucket
map during initialization, and if it turns out vbucket map is stale
the cache will be invalidated and refreshed. However, due to some
reason, the cache is not refresh correctly and no updated vbucketmap
is returned. To remove the blocker, we force vbubket replicator to
read updated vbucket map to avoid stale cached data. This may not be
the most efficient but it makes the vbucketmap always up-to-date.

Change-Id: I36ef7d69df45c3266ca051297afb4d2af3013471
Reviewed-on: http://review.couchbase.org/20365
Tested-by: buildbot <build@couchbase.com>
Tested-by: Farshid Ghods <farshid@couchbase.com>
Reviewed-by: Farshid Ghods <farshid@couchbase.com>