MB-6497: separated replication management into own gen_server
As pointed out in ticket temporary restart-type on replicators and our
existing not-ready-vbuckets handling (killing itself) do not coexist
well.

Basically, previously if there are any not ready to be replicated from
vbuckets ebucketmigrator would exit 30 seconds after it was
spawned. Before transition into temporary restart-type it would be
restarted by supervisor (note: we still had subtle races
here). Temporary restart-type caused such ebucketmigrator to be never
restarted. Part of the problem was that during rebalance we're using
update_vbucket_state requests to janitor and they're incremental. So
it would add new vbuckets, completetly dropping old.

In order to remedy this situation I've created new gen_server that
remembers desired replication topology for it's node and bucket. In
order to deal with previously unexpected and completely raceful
replicator restarts, I've moved restart-for-not-ready-vbuckets
handling into that replications manager. So requests for replication
changes (including vbucket filter change) do not race with
ebucketmigrator dying 'legitimely'.

New gen_server is placed before vbm_new_sup in it's supervisor
list. That's to make sure that if this guy crashes, vbm_sup will be
restarted as well.

Change-Id: I81e759a63e8263b4de0219c8b4f790749f908b12
Reviewed-on: http://review.couchbase.org/20492
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>