MB-6565: always remove deleted vbucket from usable set
Previously due to incorrect optimization we thought that if vbucket is
uninteresting at this point we don't have to maintain it's correct
state in usable vbuckets set, which is clearly incorrect as next
vbucket state update may ask as to add it to index and cause breakage
if we incorrectly think vbucket exists on disk.

Another similar case was when no vbucket states are set (right after
startup). In such case when we see vbucket deletion we correctly
remove it from all indexes, but we previously failed to update usable
vbuckets set.

Change-Id: I21ae921d9ca88060965a486193c319f60f83b1cb
Reviewed-on: http://review.couchbase.org/20646
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>