MB-5943: Add xdcr stats for work and commit times
Tracks and adds 2 stats to the xdcr stats UI, "secs working" and
"secs committing".  The "secs working" is the cumulative wall clock
time spent checking and writing documents, and "secs committing" is
the cumulative wall clock time spent committing (ensuring all docs
written to a target are durably written to disk) and writing the
checkpoint documents.

Change-Id: Ie82d839e846d291063e0c6af57a7dcba247c86fd
Reviewed-on: http://review.couchbase.org/20708
Reviewed-by: Damien Katz <damien@couchbase.com>
Tested-by: Damien Katz <damien@couchbase.com>
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>