MB-6586: concurent throttle state clean up
Today when vb replicator crash abnormally we may not always clean
the state correctly and as a result risk token loss during the XDC
replication, which may block all replicators if running out of the
tokens. This change-set will fix the issue and recycle the token if
a vb replicator crashed in middle of replication.

Change-Id: Idc8b78ac14a830242e8333b6928e5bf145e33c61
Reviewed-on: http://review.couchbase.org/20682
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Damien Katz <damien@couchbase.com>