MB-6637: make replication batch size pluggable
This change-set introuce a configurable parameter which allows user
to modify the size limit of batched documents in XDCR manager before
sending the packet to desinatoin cluster. The new parameter name is
"xdcr_doc_batch_size_kb" and it is defined in kilobytes. The default
value is 512, which means the size limit of batched doc is 512KB.

Change-Id: I4ba80659b43df49139d1bab1d0b67c3d049fb063
Reviewed-on: http://review.couchbase.org/20945
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>