MB-5307: automagically refresh tasks when xdcr is created/deleted
In order for our 'push' notification of xdcr task changes to work we
needed one small piece of code that handles hashing of xdcr tasks. And
previously I forgot to make it. This commit fixes the problem at
last. It will cause advertised tasks url to change when xdcr tasks is
either created or deleted causing UI to refresh list of tasks
immediately.

Change-Id: Iaffe5d405f2b45a71fe648a5c56275b31f58ffe4
Reviewed-on: http://review.couchbase.org/21043
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>