MB-6724: correctly invalidate vbucket_map_mirror cache
I.e. when any node's capi_port change is announced (i.e. due to this
node just joining in or perhaps due to config change) we need to
invalidate our cache. But this logic was broken previously causing us
not to invalidate this cache. That caused possibility of undefined
capi_url_base cached for some nodes breaking views and xdcr.

Change-Id: I8c191c92a1b5d46996eb248b08cd2452dc309bb2
Reviewed-on: http://review.couchbase.org/21111
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>