MB-6724: build just active vbuckets map in remote_clusters_info
remote_clusters_info is building full vbucket map with entries with
full CAPI urls as node identifiers. Clearly with our new policy of not
sending couchApiBase attribute for nodes without any active vbuckets
this will not work.

So instead of building full vbucket map we're only building map with
active vbuckets. So that we don't have to deal with couchApiBase-less
nodes.

Change-Id: I6994951c14f8ce48e4a8733a19d95bda3c886592
Reviewed-on: http://review.couchbase.org/21167
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>