MB-5307: eat much less memory while sending /diag
I've seen it to consume 3.7 gigs of ram when sending out diags for 8
nodes and 8 buckets. After this fix we're much better. Less then 1.5
gigs.

Main reason for memory consumption is pretty printing of large
structures which means we keep those massive strings in ram while
doing that.

So most gain is pretty printing and sending them in pieces.

Change-Id: I4b10bda39c447186d0a8a3c0f10f61f5625f2144
Reviewed-on: http://review.couchbase.org/21193
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Farshid Ghods <farshid@couchbase.com>
Tested-by: Farshid Ghods <farshid@couchbase.com>