MB-6767 Extract per ddoc and per replication stats correctly.
We try to convert stats name to an existing atom. In case such stat
does not exist we just look up corresponding binary. It implies that
any stat name is treated as known (but probably with no
samples). Apparently, it's not a big issue because we already treated
anything that could be converted to existing atom as a valid stat
name.

Change-Id: Ib45e4beb61a8d684d7f6d070bff168f1ce4cd933
Reviewed-on: http://review.couchbase.org/21367
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>