MB-6540: Replace get_disk_data with function that works on OS X
Erlang's disksup:get_disk_data() returns weird results on OS X because
it internally calls /bin/df which returns three extra columns (iNodes
Used, Free, Pct) which breaks the parse. This happens on Erlang 14B03
and 15B02.

The fix is to call /bin/df (which is what disksup does) and parse the
output correctly for OS X. For all other platforms, we continue to
call disksup for the data. Fixing this problem for OS X version 12,
but we can expand/reduce the coverage as necessary.

Change-Id: Ic040269a69e4d9ddbdf79665e6c92434d9f088b3
Reviewed-on: http://review.couchbase.org/21391
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>