MB-6809: disable flush for buckets with outgoing XDCR
I.e. because flushing of data and metadata will confuse XCDR.

Note implementation is quite naive and inherently raceful. Idealy we'd
serialize XDCR startup and flush so that it's impossible to start XDCR
after flush bucket is started. But given both features are pretty
young and immature I want to KISS instead.

Change-Id: I300d01b2e9907a90acf367c277f02bc34dbcc52c
Reviewed-on: http://review.couchbase.org/21435
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Liang Guo <liang@couchbase.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>