MB-5684: Prevent creation of buckets after a configured number
Disallow creation of buckets after total number of buckets
reaches a preset number. Currently, it defaults to 10 buckets
maximum, and this can be configured by the internal settings
dialog. The limit is enforced in the REST API. In addition,
as the REST checks trigger only after submitting a create
request in the UI, there's an additional check at UI level
as well.

Change-Id: I9590b6f808e20c563afd372fea0d107fec82f39a
Reviewed-on: http://review.couchbase.org/21467
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>