MB-5307: don't wait for second checkpoint if index pausing is off
We have internal setting to disabled index pausing as part of vbucket
movement. As part of that index pausing we also create second
checkpoint and wait for it's persistence on destination, which is
clearly not needed if indexes are not paused.

This will provide us some way to speed up rebalance without
compromising data durability.

Change-Id: Ia136068b3e7309a929316bd8252f88cb7791f60a
Reviewed-on: http://review.couchbase.org/21492
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>