MB-6869: stop 'upgrading' directory as part of config upgrade
MB-6869 is very subtle. In a bunch of places we assume directory key
is never part of dynamic config.

directory key is special because it tells you where config.dat needs
to be saved and read. And we don't want it to be replicated. Thus
weirdness.

But we found that as part of initial config upgrade that we do in
order to excercise config upgrade code we set directory key.

And that's causing it to be replicated to other nodes. Which is
causing nodes to write config.dat to wrong place.

Given that config clear is implemented through clear, save, reload
(where we reload from statically defined config path) it can now be
seen how this badness can happen. I.e. we'll read our old version of
config while clearing and saving to other node's path.

Given that controversial config upgrade is 1.6 to 1.7 which is long
not needed and not supported, it's easiest to just get rid of this
code.

NOTE: we don't remove/cleanup other config upgrade code to minimize
pre-2.0 risk.

Change-Id: Ibc59f74a104d1bdda2b4c26677097e0c4d28b4fa
Reviewed-on: http://review.couchbase.org/21495
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>