MB-6863: sync old dbdir to couch's db and index dir config keys
Since long time ago 2.0 is using couch config to store this two
settings. That's because we need to sync them to couch config anyways
and having 2 copies is unhelpful.

But that raises upgrade question. Particular somebody running 1.8.1
may have custom db dir defined. That would be stored in per-node
memcached key in ns_config. And date conversion code is assuming we'll
keep this path.

We deal with that as follows. As part of init of cb_config_couch_sync
we check if ns_config has dbdir and if it does we sync it to couch and
_delete_ from ns_config. So that next time we won't overwrite it.

Both db and ix dirs are set to old dbdir setting in ns_config.

NOTE: normally when couchdb's db or ix dir is updated we restart
everything. I.e. see commit
24132400404b31d0458e66e39708a42ebf7dc5c0. In this case full restart
can be avoided because cb_config_couch_sync is part of ns_config_sup
which is started before any ns_server's services that may depend on
couch.

Change-Id: I6c278bf258d82a8059becf234d46cbba3aba4548
Reviewed-on: http://review.couchbase.org/21504
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>