MB-6860: Get signatures without opening group
For the spatial index cleanup we need to get the signatures
of the groups. Previously we opened the group to get it,
that lead to creating spatial groups even for design document
that don't contain any spatial indexes.

This commit also makes it more robust like removing all .spatial
files when there are no design documents at all.

This commit depends on: http://review.couchbase.org/21591

Change-Id: I54189ef1869806060479e0b1add9c72ec0065a35
Reviewed-on: http://review.couchbase.org/21593
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>