MB-6889 Be consistent on data path in cluster_run.
With recent change that syncs memcached dbdir to couch_config, on the
first invocation of the server the cluster would use data path
inherited from dbdir, on the second invocation cluster_run would
overwrite the paths in configs with something different from
dbdir. Thus all the data would be lost. The solution is just to be
consistent on what data path we use.

Change-Id: Icf1df3af94b47388d575ea0324bd5686a09d1acf
Reviewed-on: http://review.couchbase.org/21565
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>