MB-6919: expose more XDCR stats to UI
This change-set provides more XDCR stats on UI to help QE, performance
team and customers understand the XDCR operations better. We add 3 new
outbound operation stats, namely, 1) data replicated, which is amount
of data XDCR have replicated to the remote data center through inter-
cluster network; 2) active vb replications, which shows the number of
ongoing active replications to the remote clusters; 3) waiting vb
replications, which is the number of vb replications in the waiting
pool to wait for turn to start replication. These stats will be useful
to performance monitoring and diagnosis.

Change-Id: If9d42cde5dd715404f0b1bcf26ea143493ae9f75
Reviewed-on: http://review.couchbase.org/21613
Reviewed-by: Junyi Xie <junyi.couchbase@gmail.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>