MB-6918 Remove memcached dbdir from default config.
It's not needed in 2.0. And with recent change that synchronizes it to
database_dir and view_index_dir it may be confused with genuine dbdir
left from the older version of couchbase server. Particularly, if some
node is removed from the cluster its config gets cleared and dbdir
will again become part of memcached config. It means if this node had
a custom data path it will be overwritten on the next restart.

Change-Id: I158309323506863790044285a9fb04cd2d3718da
Reviewed-on: http://review.couchbase.org/21696
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>