MB-6494: handle empty ready vbuckets set in ebucketmigrator
We cannot send empty vbucket filter to ep-engine. If we do, ep-engine
will send all vbuckets, instead of none.

Thus when all vbuckets are not ready to be replicated from, we have to
nothing. We cannot die, because it will cause crashes above.

We had such behavior previously, but apprently recent code for
handling new-style vbucket filter change or something else removed it.

Note, there's also effect on vbucket filter change. When we hit this
condition, we force old-style vbucket filter change. That's because
there's no existing tap producer with our tap name.

Because no tap stream is going to be used while we do nothing (waiting
restart message), we kill tap stream with our name (if any).

Change-Id: Ic868a059fddf5055290b783882031453e3fb2ceb
Reviewed-on: http://review.couchbase.org/21733
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>