MB-7014: anticipate EXIT from port when waiting for shutdown
In logs of MB-7014 we timeout waiting for port shutdown because we
didn't receive exit_status from port, but only EXIT with epipe. Which
we ignored (and logged). Thus 60 seconds timeout on waiting for child
was hit, when in fact child was dead long time ago.

It's quite sad that epipe leads to loss of exit status. Basically we
lose exit status if before receiving exit status we try to send
anything to port. But that's life, and I see nothing that we can do
here.

Change-Id: I3af66bfabe77ae9654dfe880bc0387b6fbfe23ee
Reviewed-on: http://review.couchbase.org/21974
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>