MB-7000 Fake per node memcached config vclock.
We fake it in the way that it appears to be edited by some node
'_'. We need this because in 1.8 this key can be changed more times
than in clean 2.0. It means that its vclock will have greater counters
and will descend 2.0's vclock. So if the node was part of the cluster
with old version of software, then was removed from the cluster and
the server was completely removed and then 2.0 was installed, after it
gets added back the old value of memcached configuration will override
the new one. A change from a fake node will cause vclock conflict and
new value will be preferred because of the newer timestamp.

Change-Id: I5810e5ddbbec426149f3801d34af71826afeb28c
Reviewed-on: http://review.couchbase.org/21999
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Sriram Melkote <siri@couchbase.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>