MB-7040: use cluster_compat_mode to send out clusterCompatibility
Because entire cluster runs in same cluster compat there's no need to
rely on heartbeat for that.

Plus taking it from heartbeat has risk of taking stale version of this
information. Which is what happens in MB-7040. When node is ejected it
initially runs in 1.8 compat mode (due to ns_config upgrade issues we
run into), but it'll immediately upgrade itself to 2.0 mode when
mb_master will start ns_orchestrator. But because it happens after
ns_doctor and ns_heart are started, node actually has stale node info
about itself in ns_doctor.

Change-Id: I2b18420b3babeb36c0da118b3bf8543d19590c2f
Reviewed-on: http://review.couchbase.org/22068
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>