MB-5308: display index task with higher priority
On views UI when both compaction and indexing is running we have to
pick which progress to display. Our previous decision was to pick
compaction because it was supposed to be faster.

But what I saw in practice is that compaction runs quite often and
quite long when index updater is running. Moreover I've seen 100%
compaction progress for quite long time.

It appears that seeing indexing being X% done rather than compaction
being 100% done is more useful.

Change-Id: Ie5e34058af25e9235b88c7e2886780b07f87d508
Reviewed-on: http://review.couchbase.org/22121
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>
Tested-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>