MB-7187: avoid saving junk into design docs
After we've separated doc metadata and made ddocs inside JS have .json
and .meta attributes, we failed to clone ddoc properly for displaying
on you.

I.e. _.clone was shallow clone which caused json field to actually be
shared and which caused extra fields that we add just for UI templates
to be saved.

Change-Id: I63599faa6d65f14ce52398e68e388f51fe13e2d8
Reviewed-on: http://review.couchbase.org/22386
Reviewed-by: Farshid Ghods <farshid@couchbase.com>
Tested-by: Farshid Ghods <farshid@couchbase.com>