MB-7132: change XDCR stats name from docs to mutations
Today the word "docs" in outbound XDCR stats section indeed refers
to the number of mutations. Say, if one docusments is updated twice,
we will see 2 "docs to replicate" in the stats section. This is a bit
misleading since there is only 1 doc created in system. We now use
"mutations to replicate" which looks better than "docs to replicate".

Change-Id: Ia1b13e8f76cfc2727e9d73b3ae8e40707b5059e5
Reviewed-on: http://review.couchbase.org/22381
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>