MB-7133: change default batch size and log improvement
Today during XDCR, the batch is limited by both the number of docs
and size of batch packet. The former is set to 500 docs. But the
latter is only 512K, which may nullify the number of docs limit
if the averge doc size is more than 1KB. By Fashid, it is reasonable
to assume 4KB doc in the workload, accordingly the size limit is set
to 2048KB.

Additionally, in this change-set we 1) make the concurrent reps per
bucket a ns_server configurable parameter, consitent with other XDCR
parameters. Note the env parameter name is not changed; 2) dump the
default value of all XDCR parameters when the replication manager
starts, for easy debug and performance diag.

Change-Id: If2cd6db1ce660afdb54179790072ba97e5005661
Reviewed-on: http://review.couchbase.org/22551
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Junyi Xie <junyi.couchbase@gmail.com>
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>