MB-7523: start backfill of next vbuckets as soon previous is done
As part of MB-6726 we've implemented ability to serialize just initial
phase of vbucket move where bulk of work is performed by
ep-engine. I.e. it's still not advisable to overload ep-engine with
multiple incoming or outgoing tap backfills. But it's great to have
second phase (mostly indexing) to be parallelized.

But MB-6726 ends that serial phase only when first checkpoint we're
waiting for hits disk on future master. And there's no need for this
to be affected by latency of ram -> disk.

Especially given that due to now famous "fsync" flaw in couchstore
design we see very severe latency between getting data into ram on
destination node and putting it on disk there, particularly when
there's even minimal mutations load during this time. I.e. because
ep-engine flusher has to independently fsync and commit a bunch of
vbuckets, while only one "our" vbucket is what's being waited for.

With that patch we're removing that unneeded latency "dependency", and
I've observed dramatic improvement of rebalance run time under load.

Change-Id: Ib4f52b1def2efceea1cfb41daadcc5cc9750d518
Reviewed-on: http://review.couchbase.org/24067
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Artamonau <aliaksiej.artamonau@gmail.com>